refactor out libwriters, fix references to Timestamp, Timedelta
