+++ pandas/io/stata.py
@@ -16,8 +16,9 @@ import sys
 
 import numpy as np
 from dateutil.relativedelta import relativedelta
-from pandas._libs.lib import max_len_string_array, infer_dtype
+from pandas._libs.lib import infer_dtype
 from pandas._libs.tslib import NaT, Timestamp
+from pandas._libs.writers import max_len_string_array
 
 import pandas as pd
 from pandas import compat, to_timedelta, to_datetime, isna, DatetimeIndex
